Update last state.
