+++ ontology/country_ontology.ttl
@@ -2107,7 +2107,7 @@ cx:City a owl:Class ;
 
 <https://raw.githubusercontent.com/catenax-ng/product-knowledge/main/ontology/country_ontology.ttl> a owl:Ontology ;
     dc:creator "Zazralt Magic" ;
-    dc:date "2022-11-28"^^xsd:date ;
+    dc:date "2023-02-11"^^xsd:date ;
     dc:description "Ontology with countries. Which countires are in the world? What are their ISO codes?" ;
     dc:title "Country Ontology" ;
     owl:versionInfo "0.0.1" .
@@ -2116,8 +2116,11 @@ cx:Country a owl:Class ;
     rdfs:isDefinedBy <https://raw.githubusercontent.com/catenax-ng/product-knowledge/main/ontology/country_ontology.ttl> ;
     rdfs:seeAlso <https://en.wikipedia.org/wiki/Country> ;
     rdfs:subClassOf [ a owl:Restriction ;
-            owl:onProperty cx:wikidata ;
-            owl:someValuesFrom xsd:anyURI ],
+            owl:onProperty cx:isLocatedIn ;
+            owl:someValuesFrom cx:Continent ],
+        [ a owl:Restriction ;
+            owl:onProperty cx:countryName ;
+            owl:someValuesFrom xsd:string ],
         [ a owl:Restriction ;
             owl:onProperty cx:countryNumber ;
             owl:someValuesFrom xsd:integer ],
@@ -2125,17 +2128,14 @@ cx:Country a owl:Class ;
             owl:onProperty cx:hasCapitalCity ;
             owl:someValuesFrom cx:CapitalCity ],
         [ a owl:Restriction ;
-            owl:onProperty cx:countryName ;
-            owl:someValuesFrom xsd:string ],
-        [ a owl:Restriction ;
             owl:onProperty cx:countryCode3 ;
             owl:someValuesFrom xsd:string ],
         [ a owl:Restriction ;
-            owl:onProperty cx:isLocatedIn ;
-            owl:someValuesFrom cx:Continent ],
-        [ a owl:Restriction ;
             owl:onProperty cx:countryCode2 ;
             owl:someValuesFrom xsd:string ],
+        [ a owl:Restriction ;
+            owl:onProperty cx:wikidata ;
+            owl:someValuesFrom xsd:anyURI ],
         cx:Place ;
     skos:altLabel "countries"@en,
         "state"@en,
