+++ ontology/ontology_mapping/bamm_ess_incident_1_0_0_mapping.csv
@@ -1,0 +1,68 @@
+type,map,class,table_name,table_type,table_path,attribute,subject,column,data_type,relation,foreign_key,target_map,primary_key
+# prefix,cx,,,,,,,,,,,,
+# namespace,https://raw.githubusercontent.com/catenax-ng/product-knowledge/main/ontology/,,,,,,,,,,,,
+# title,,,,,,,,,,,,,
+# version,0.0.1,,,,,,,,,,,,
+# author,Zazralt Magic,,,,,,,,,,,,
+# contributor,,,,,,,,,,,,,
+# description,,,,,,,,,,,,,
+# import,,,,,,,,,,,,,
+# table_type,,,,"SQL2008, Parquet, CSV, XPath, JSONPath",,,,,,,,,
+,,,,,,,,,,,,,
+attribute,cx:BatchPartTripleMap,cx:BatchPart,,,,cx:batchNumber,,,xsd:string,,,,
+attribute,cx:BatchPartTripleMap,cx:BatchPart,,,,cx:batchProductionDate,,,xsd:date,,,,
+attribute,cx:BatchPartTripleMap,cx:BatchPart,,,,cx:batchProductionDateTime,,,xsd:dateTime,,,,
+relation,cx:BatchPartTripleMap,cx:BatchPart,,,,,,,,cx:isAffectedIn,,cx:IncidentReportTripleMap,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentCause,,,xsd:string,,,,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentDamage,,,xsd:string,,,,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentDateTime,,,xsd:dateTime,,,,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentDescription,,,xsd:string,,,,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentDuration,,,xsd:string,,,,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentHazard,,,xsd:string,,,,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentImpact,,,xsd:string,,,,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentInjury,,,xsd:string,,,,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentLocation,,,xsd:string,,,,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentPrevention,,,xsd:string,,,,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentRisk,,,xsd:string,,,,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentSeverity,,,xsd:string,,,,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentTitle,,,xsd:string,,,,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentType,,,xsd:string,,,,
+attribute,cx:IncidentTripleMap,cx:Incident,,,,cx:incidentUrgency,,,xsd:string,,,,
+relation,cx:IncidentTripleMap,cx:Incident,,,,,,,,cx:affects,,cx:IncidentAffectedCompanyTripleMap,
+relation,cx:IncidentTripleMap,cx:Incident,,,,,,,,cx:affects,,cx:IncidentAffectedPersonTripleMap,
+relation,cx:IncidentTripleMap,cx:Incident,,,,,,,,cx:isCausedBy,,cx:IncidentCausingPersonTripleMap,
+relation,cx:IncidentTripleMap,cx:Incident,,,,,,,,cx:isCausedBy,,cx:IncidentCausingCompanyTripleMap,
+relation,cx:IncidentTripleMap,cx:Incident,,,,,,,,cx:isLocatedIn,,cx:PlaceTripleMap,
+relation,cx:IncidentTripleMap,cx:Incident,,,,,,,,cx:isLocatedIn,,cx:StreetAddressTripleMap,
+relation,cx:IncidentTripleMap,cx:Incident,,,,,,,,cx:isRecordedBy,,cx:IncidentReportTripleMap,
+relation,cx:IncidentTripleMap,cx:Incident,,,,,,,,cx:isWitnessedBy,,cx:IncidentWitnessTripleMap,
+relation,cx:IncidentAffectedCompanyTripleMap,cx:IncidentAffectedCompany,,,,,,,,cx:hasContact,,cx:BusinessContactTripleMap,
+relation,cx:IncidentAffectedCompanyTripleMap,cx:IncidentAffectedCompany,,,,,,,,cx:isAffectedBy,,cx:IncidentTripleMap,
+relation,cx:IncidentAffectedCompanyTripleMap,cx:IncidentAffectedCompany,,,,,,,,cx:isLocatedIn,,cx:BusinessAddressTripleMap,
+attribute,cx:IncidentReportTripleMap,cx:IncidentReport,,,,cx:incidentReportAnonymousFlag,,,xsd:boolean,,,,
+attribute,cx:IncidentReportTripleMap,cx:IncidentReport,,,,cx:incidentReportDateTime,,,xsd:dateTime,,,,
+attribute,cx:IncidentReportTripleMap,cx:IncidentReport,,,,cx:incidentReportEvidenceAttachment,,,xsd:anyURI,,,,
+attribute,cx:IncidentReportTripleMap,cx:IncidentReport,,,,cx:incidentReportId,,,xsd:string,,,,
+attribute,cx:IncidentReportTripleMap,cx:IncidentReport,,,,cx:incidentReportStatus,,,xsd:string,,,,
+attribute,cx:IncidentReportTripleMap,cx:IncidentReport,,,,cx:notifiedAuthority,,,xsd:string,,,,
+relation,cx:IncidentReportTripleMap,cx:IncidentReport,,,,,,,,cx:isCreatedBy,,cx:IncidentReporterTripleMap,
+relation,cx:IncidentReportTripleMap,cx:IncidentReport,,,,,,,,cx:isReceivedBy,,cx:OrganisationTripleMap,
+relation,cx:IncidentReportTripleMap,cx:IncidentReport,,,,,,,,cx:isReceivedBy,,cx:CompanyTripleMap,
+attribute,cx:IncidentReporterTripleMap,cx:IncidentReporter,,,,cx:incidentReporterCompany,,,xsd:string,,,,
+attribute,cx:IncidentReporterTripleMap,cx:IncidentReporter,,,,cx:incidentReporterId,,,xsd:string,,,,
+attribute,cx:IncidentReporterTripleMap,cx:IncidentReporter,,,,cx:incidentReporterName,,,xsd:string,,,,
+attribute,cx:IncidentReporterTripleMap,cx:IncidentReporter,,,,cx:incidentReporterSignature,,,xsd:string,,,,
+relation,cx:IncidentReporterTripleMap,cx:IncidentReporter,,,,,,,,cx:hasContact,,cx:ContactTripleMap,
+relation,cx:IncidentReporterTripleMap,cx:IncidentReporter,,,,,,,,cx:isLocatedIn,,cx:StreetAddressTripleMap,
+attribute,cx:PartTripleMap,cx:Part,,,,cx:partId,,,xsd:string,,,,
+attribute,cx:PartTripleMap,cx:Part,,,,cx:partName,,,xsd:string,,,,
+attribute,cx:PartTripleMap,cx:Part,,,,cx:partProductionDate,,,xsd:date,,,,
+attribute,cx:PartTripleMap,cx:Part,,,,cx:partProductionDateTime,,,xsd:dateTime,,,,
+attribute,cx:PartTripleMap,cx:Part,,,,cx:partSeries,,,xsd:string,,,,
+relation,cx:PartTripleMap,cx:Part,,,,,,,,cx:hasError,,cx:VehicleErrorTripleMap,
+relation,cx:PartTripleMap,cx:Part,,,,,,,,cx:hasLoadSpectrum,,cx:LoadSpectrumTripleMap,
+relation,cx:PartTripleMap,cx:Part,,,,,,,,cx:isAffectedIn,,cx:IncidentReportTripleMap,
+relation,cx:PartTripleMap,cx:Part,,,,,,,,cx:isPartOf,,cx:PartTripleMap,
+relation,cx:PartTripleMap,cx:Part,,,,,,,,cx:isPartOf,,cx:VehicleTripleMap,
+relation,cx:PartTripleMap,cx:Part,,,,,,,,cx:isProducedBy,,cx:PartManufacturerTripleMap,
+relation,cx:PartTripleMap,cx:Part,,,,,,,,cx:isReasonOf,,cx:PossibleCauseTripleMap,
