+++ ontology/ontology_mapping/bamm_ess_incident_1_0_0_mapping.ttl
@@ -1,0 +1,21 @@
+@prefix cx: <https://raw.githubusercontent.com/catenax-ng/product-knowledge/main/ontology/cx_ontology.ttl#> .
+@prefix rml: <http://semweb.mmlab.be/ns/rml#> .
+@prefix rr: <http://www.w3.org/ns/r2rml#> .
+@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .
+
+cx:IncidentTripleMap a rr:TriplesMap ;
+    rr:logicalTable [ rml:referenceFormulation rr:SQL2008 ;
+            rml:source "json_sample"@en ;
+            rr:tableName "essIncident"@en ] ;
+    rr:predicateObjectMap [ rr:objectMap [ rr:column "essIncidentInformation.incidentDate" ;
+                    rr:datatype xsd:dateTime ] ;
+            rr:predicate cx:incidentDateTime ],
+        [ rr:objectMap [ rr:column "essIncidentInformation.incidentHeadline" ;
+                    rr:datatype xsd:string ] ;
+            rr:predicate cx:incidentTitle ] ;
+    rr:subjectMap [ rr:class cx:Incident ;
+            rr:column """    essIncidentInformation.incidentDate
+essIncidentInformation.incidentHeadline""" ;
+            rr:template """https://raw.githubusercontent.com/catenax-ng/product-knowledge/main/ontology/cx_ontology.ttl#Incident_{    essIncidentInformation.incidentDate
+essIncidentInformation.incidentHeadline}""" ] .
+
